requirements.txt
pyocr
pillow
